Move encodings to a private submodule
