+++ singer_sdk/_singerlib/_encoding/__init__.py
