+++ singer_sdk/_singerlib/_encoding/base.py
