+++ singer_sdk/_singerlib/_encoding/simple.py
